refactor(core): ♻️ extract performance mode and service worker from skin.js
