+++ resources/skins.citizen.scripts/skin.js
@@ -3,9 +3,10 @@
  */
 function deferredTasks() {
 	const speculationRules = require( './speculationRules.js' );
+	const serviceWorker = require( './serviceWorker.js' );
 
 	speculationRules.init();
-	registerServiceWorker();
+	serviceWorker.register();
 
 	window.addEventListener( 'beforeunload', () => {
 		// Set up loading indicator
@@ -18,31 +19,6 @@ function deferredTasks() {
 	} );
 
 	document.documentElement.classList.add( 'citizen-animations-ready' );
-}
-
-/**
- * Register service worker
- *
- * @return {void}
- */
-function registerServiceWorker() {
-	const scriptPath = mw.config.get( 'wgScriptPath' );
-	// Only allow serviceWorker when the scriptPath is at root because of its scope
-	// I can't figure out how to add the Service-Worker-Allowed HTTP header
-	// to change the default scope
-	if ( scriptPath !== '' ) {
-		return;
-	}
-
-	if ( 'serviceWorker' in navigator ) {
-		const SW_MODULE_NAME = 'skins.citizen.serviceWorker',
-			version = mw.loader.moduleRegistry[ SW_MODULE_NAME ].version,
-			// HACK: Faking a RL link
-			swUrl = scriptPath +
-				'/load.php?modules=' + SW_MODULE_NAME +
-				'&only=scripts&raw=true&skin=citizen&version=' + version;
-		navigator.serviceWorker.register( swUrl, { scope: '/' } );
-	}
 }
 
 /**
@@ -80,41 +56,6 @@ function initPreferences() {
 }
 
 /**
- * Turn off performance mode if GPU acceleration is available
- * Skip the check if a user preference is already set.
- *
- * @return {void}
- */
-function initPerformanceMode() {
-	const prefName = 'citizen-feature-performance-mode-clientpref-';
-	const clientPrefs = localStorage.getItem( 'mwclientpreferences' );
-
-	if ( clientPrefs && ( clientPrefs.includes( prefName + '0' ) || clientPrefs.includes( prefName + '1' ) ) ) {
-		return;
-	}
-
-	const canvas = document.createElement( 'canvas' );
-	const contextNames = [ 'webgl', 'experimental-webgl', 'webgl2' ];
-	const hasGpu = contextNames.some( ( name ) => {
-		try {
-			const gl = canvas.getContext( name );
-			return !!( gl && typeof gl.getParameter === 'function' );
-		} catch ( e ) {
-			return false;
-		}
-	} );
-
-	if ( !hasGpu ) {
-		// Performance mode ON is default, so we don't need to do anything
-		return;
-	}
-
-	document.documentElement.classList.remove( prefName + '1' );
-	document.documentElement.classList.add( prefName + '0' );
-	localStorage.setItem( 'mwclientpreferences', clientPrefs ? `${ clientPrefs },${ prefName }0` : `${ prefName }0` );
-}
-
-/**
  * @param {Window} window
  * @return {void}
  */
@@ -126,7 +67,8 @@ function main( window ) {
 		dropdown = require( './dropdown.js' ),
 		lastModified = require( './lastModified.js' ),
 		share = require( './share.js' ),
-		setupObservers = require( './setupObservers.js' );
+		setupObservers = require( './setupObservers.js' ),
+		performance = require( './performance.js' );
 
 	search.init( window );
 	echo();
@@ -134,7 +76,7 @@ function main( window ) {
 	dropdown.init();
 	lastModified.init();
 	share.init();
-	initPerformanceMode();
+	performance.init();
 
 	mw.hook( 'wikipage.content' ).add( ( content ) => {
 		// content is a jQuery object
